feat(meetings): entity router with crud + reads + participants sub-routers
Replace flat meetings.router.ts with directory-based entity router using
createEntityRouter factory. CRUD operations use createCrudRouter with
lifecycle callbacks. Reads (list, getByIdWithJoins, getInternalUsers)
and participants (getParticipants, manageParticipants) extracted into
dedicated sub-routers. All db.update(meetings).set({ownerId}) calls
replaced with meetingCrud.update() via dalVerifySuccess.
